mathics.eval.string.to_regex -> mathics.core.convert.regex.to_regex
